requirements.txt
Appium-Python-Client==0.26
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
ddt==1.2.0
et-xmlfile==1.0.1
Flask==1.1.1
Flask-SocketIO==4.2.1
future==0.17.1
idna==2.8
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
MarkupSafe==1.1.1
openpyxl==2.6.2
Pillow==7.0.0
pycryptodome==3.6.6
python-engineio==3.11.2
python-socketio==4.3.1
requests==2.22.0
selenium==3.0.1
six==1.14.0
soupsieve==2.0
urllib3==1.25.3
Werkzeug==0.15.5
